Add Federation Plugin support
The output format for federated queues and exchanges differs from
normal ones, so the modeler has been updated to handle that format.
Some additional safety has been added, such as handling the default '/'
exchange better.

There is also now a bool property on queues and exchanges to indicate whether
they are federated or not.

Fixes ZEN-20898